Moved folder names to `swervedrive`. Created AutoBalanceCommand from @frc3512's code.
