+++ docs/json/controllerproperties.md
@@ -3,9 +3,9 @@
 The Swerve Controller stores configuration options relating it how the swerve drive works during
 autonomous and drive modes which set the heading of the robot based off a joystick.
 The JSON files maps 1:1
-with [`ControllerPropertiesJson.java`](../../src/main/java/frc/robot/subsystems/swervedrive2/swervelib/parser/json/ControllerPropertiesJson.java)
+with [`ControllerPropertiesJson.java`](../../src/main/java/frc/robot/subsystems/swervedrive/swervelib/parser/json/ControllerPropertiesJson.java)
 which creates
-a [`SwerveControllerConfiguration`](../../src/main/java/frc/robot/subsystems/swervedrive2/swervelib/parser/SwerveControllerConfiguration.java).
+a [`SwerveControllerConfiguration`](../../src/main/java/frc/robot/subsystems/swervedrive/swervelib/parser/SwerveControllerConfiguration.java).
 The values within here are **EXTREMELY** important to autonomous because the robot heading PID needs
 to be tuned correctly in order for Autonomous functions to work.
