+++ docs/json/pidfproperties.md
@@ -2,7 +2,7 @@
 
 This file configures the PIDF values with integral zone and maximum output of the drive and motor
 modules for every swerve module. It maps 1:1
-to [`PIDFPropertiesJson.java`](../../src/main/java/frc/robot/subsystems/swervedrive2/swervelib/parser/json/PIDFPropertiesJson.java)
+to [`PIDFPropertiesJson.java`](../../src/main/java/frc/robot/subsystems/swervedrive/swervelib/parser/json/PIDFPropertiesJson.java)
 and is used while initializing the `SwerveDriveConfiguration` object.
 
 # Fields
