+++ src/main/java/frc/robot/commands/swervedrive/auto/FollowTrajectory.java
@@ -1,10 +1,10 @@
-package frc.robot.commands.swervedrive2.auto;
+package frc.robot.commands.swervedrive.auto;
 
 import com.pathplanner.lib.PathPlannerTrajectory;
 import com.pathplanner.lib.commands.PPSwerveControllerCommand;
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
 import frc.robot.Constants.Auton;
-import frc.robot.subsystems.swervedrive2.SwerveSubsystem;
+import frc.robot.subsystems.swervedrive.SwerveSubsystem;
 
 public class FollowTrajectory extends SequentialCommandGroup
 {
