+++ src/main/java/frc/robot/commands/swervedrive/drivebase/AbsoluteFieldDrive.java
@@ -2,14 +2,15 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands.swervedrive2.drivebase;
+package frc.robot.commands.swervedrive.drivebase;
 
+import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.math.geometry.Translation2d;
 import edu.wpi.first.math.kinematics.ChassisSpeeds;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.Constants;
-import frc.robot.subsystems.swervedrive2.SwerveSubsystem;
+import frc.robot.subsystems.swervedrive.SwerveSubsystem;
 import java.util.List;
 import java.util.function.DoubleSupplier;
 import swervelib.SwerveController;
@@ -18,12 +19,11 @@ import swervelib.math.SwerveMath;
 /**
  * An example command that uses an example subsystem.
  */
-public class AbsoluteDrive extends CommandBase
+public class AbsoluteFieldDrive extends CommandBase
 {
 
   private final SwerveSubsystem swerve;
-  private final DoubleSupplier  vX, vY;
-  private final DoubleSupplier headingHorizontal, headingVertical;
+  private final DoubleSupplier  vX, vY, heading;
   private final boolean isOpenLoop;
 
   /**
@@ -32,28 +32,21 @@ public class AbsoluteDrive extends CommandBase
    * coordinates from which the robot's angle will be derived— they will be converted to a polar angle, which the robot
    * will rotate to.
    *
-   * @param swerve            The swerve drivebase subsystem.
-   * @param vX                DoubleSupplier that supplies the x-translation joystick input.  Should be in the range -1
-   *                          to 1 with deadband already accounted for.  Positive X is away from the alliance wall.
-   * @param vY                DoubleSupplier that supplies the y-translation joystick input.  Should be in the range -1
-   *                          to 1 with deadband already accounted for.  Positive Y is towards the left wall when
-   *                          looking through the driver station glass.
-   * @param headingHorizontal DoubleSupplier that supplies the horizontal component of the robot's heading angle. In the
-   *                          robot coordinate system, this is along the same axis as vY. Should range from -1 to 1 with
-   *                          no deadband.  Positive is towards the left wall when looking through the driver station
-   *                          glass.
-   * @param headingVertical   DoubleSupplier that supplies the vertical component of the robot's heading angle.  In the
-   *                          robot coordinate system, this is along the same axis as vX.  Should range from -1 to 1
-   *                          with no deadband. Positive is away from the alliance wall.
+   * @param swerve  The swerve drivebase subsystem.
+   * @param vX      DoubleSupplier that supplies the x-translation joystick input.  Should be in the range -1 to 1 with
+   *                deadband already accounted for.  Positive X is away from the alliance wall.
+   * @param vY      DoubleSupplier that supplies the y-translation joystick input.  Should be in the range -1 to 1 with
+   *                deadband already accounted for.  Positive Y is towards the left wall when looking through the driver
+   *                station glass.
+   * @param heading DoubleSupplier that supplies the robot's heading angle.
    */
-  public AbsoluteDrive(SwerveSubsystem swerve, DoubleSupplier vX, DoubleSupplier vY, DoubleSupplier headingHorizontal,
-                       DoubleSupplier headingVertical, boolean isOpenLoop)
+  public AbsoluteFieldDrive(SwerveSubsystem swerve, DoubleSupplier vX, DoubleSupplier vY,
+                            DoubleSupplier heading, boolean isOpenLoop)
   {
     this.swerve = swerve;
     this.vX = vX;
     this.vY = vY;
-    this.headingHorizontal = headingHorizontal;
-    this.headingVertical = headingVertical;
+    this.heading = heading;
     this.isOpenLoop = isOpenLoop;
 
     addRequirements(swerve);
@@ -72,8 +65,7 @@ public class AbsoluteDrive extends CommandBase
     // Get the desired chassis speeds based on a 2 joystick module.
 
     ChassisSpeeds desiredSpeeds = swerve.getTargetSpeeds(vX.getAsDouble(), vY.getAsDouble(),
-                                                         headingHorizontal.getAsDouble(),
-                                                         headingVertical.getAsDouble());
+                                                         new Rotation2d(heading.getAsDouble() * Math.PI));
 
     // Limit velocity to prevent tippy
     Translation2d translation = SwerveController.getTranslation2d(desiredSpeeds);
